DCS-778 Add an appointment location finder end-point that is simpler for the BVL front-end to use.
